Adyen style, gallery layout
The Adyen submit button and the surrounding form has been styled
according to the latest feedback.

The image gallery fills out the entire container width, at the cost
of some images being zoomed in.

Nav logo is now a link to /, and the desktop nav width resembles the
desktop content width.