Payment integration, vol.1
It is now possible to go through the MobilePay checkout and get
redirected back to the site.

The upcoming commit will need to handle the various payment statuses,
as well as the process of checking for them.

Form has been adjusted to this functionality, including loading, and
error handling.

Changed error handling in get-current-user-amount function to be at
least a bit more transparent.