Implement feedback
Including various visual improvements to more closely match the feum
website, as well as another state for the confirmation screen.

A new field added for the event's artists, which can hold a music page
link, and should not be required in the CMS alongside the label field.

Navigation simplified.

Main font switched to Helvetica.